+++ test/test_repos.rb
@@ -7,7 +7,6 @@ module TestRepos
     `rm -rf #{path}`
   end
 
-
   def git_projects_path(project_path)
     path = Dir.pwd+project_path
     `mkdir -p #{path}` unless File.directory?(path)